ich_museum/requirements/base.txt
django==4.2.*
djangorestframework==3.14.*
djangorestframework-simplejwt==5.3.*
django-cors-headers==4.3.*
django-redis==5.4.*
redis==5.0.*
psycopg2-binary==2.9.*
drf-spectacular==0.27.*
python-dotenv==1.0.*
elasticsearch==8.11.*
